Implemented additional post process usage in `SimualtionSet.run`. Resolves issue 5.
The `run` method now has additional input parameters
`run_post_process_files` and `additional_keys`. In contrast to the
implementation in `solve_single_simulation`, a merged project file is
created, since a subsequent execution of the post processes would have
been to hard to implement at this time. The additional keys are passed
to the actual solve as well, which should not cause to much trouble.

The notebook `Post process management` demonstrates the new features but
needs documentation!