Restructuring of antenna extension finished.
The antenna extension now provides the class `FarFieldEvaluation` which
takes a `core.Simulation` instance as input parameter in the
constructor. The user calls the `analyze_far_field`-method, while the
evaluator takes care of solving the simulation if necessary.

This needs to be checked using simulation sets.

There are still issues if `direction` is not None.

The 'Extensions explained - antenna' notebook demonstrates the new use
case but still needs some documentation.